refactor: extract paste extmark logic into usePasteExtmarks hook
